# openapi.json
{"openapi":"3.1.0","info":{"title":"Curriculum Management API","description":"API untuk manajemen kurikulum, CPL, dan mata kuliah","version":"1.0.0"},"paths":{"/":{"get":{"summary":"Main","operationId":"main__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/auth/login":{"post":{"tags":["Authentication"],"summary":"Login User","description":"Autentikasi user dan mendapatkan JWT access token","operationId":"login_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"JWT access token untuk autentikasi dan otorisasi endpoint lain","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"401":{"description":"Unauthorized - Invalid credentials or token"},"400":{"description":"Bad Request"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/me":{"get":{"tags":["Authentication"],"summary":"Get Current User Info","description":"Mengambil informasi user yang sedang login berdasarkan JWT token","operationId":"get_current_user_info_auth_me_get","responses":{"200":{"description":"Data lengkap user yang sedang terautentikasi","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"401":{"description":"Unauthorized - Invalid credentials or token"},"400":{"description":"Bad Request"}},"security":[{"HTTPBearer":[]}]}},"/auth/logout":{"post":{"tags":["Authentication"],"summary":"Logout User","description":"Logout user dengan cara memasukkan token ke blacklist","operationId":"logout_auth_logout_post","responses":{"200":{"description":"Konfirmasi logout berhasil","content":{"application/json":{"schema":{}}}},"401":{"description":"Unauthorized - Invalid credentials or token"},"400":{"description":"Bad Request"}},"security":[{"HTTPBearer":[]}]}},"/auth/register":{"post":{"tags":["Authentication"],"summary":"Register User Baru","description":"Mendaftarkan user baru ke sistem, hanya bisa dilakukan kadep","operationId":"register_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"201":{"description":"Data user yang berhasil didaftarkan","content":{"application/json":{"schema":{}}}},"401":{"description":"Unauthorized - Invalid credentials or token"},"400":{"description":"Bad Request"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/kurikulum/":{"get":{"tags":["kurikulum"],"summary":"Daftar Semua Kurikulum","description":"Mengambil daftar lengkap semua kurikulum yang ada di sistem","operationId":"get_all_kurikulum__get","responses":{"200":{"description":"Total dan daftar kurikulum","content":{"application/json":{"schema":{}}}},"404":{"description":"Tidak ditemukan"}},"security":[{"HTTPBearer":[]}]},"post":{"tags":["kurikulum"],"summary":"Tambah Kurikulum Baru","description":"Menambahkan kurikulum baru ke dalam sistem","operationId":"create_kurikulum_kurikulum__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/KurikulumCreate"}}},"required":true},"responses":{"201":{"description":"Data kurikulum yang berhasil ditambahkan","content":{"application/json":{"schema":{}}}},"404":{"description":"Tidak ditemukan"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/kurikulum/{id_kurikulum}":{"patch":{"tags":["kurikulum"],"summary":"Update Kurikulum","description":"Mengupdate informasi kurikulum yang sudah ada","operationId":"update_kurikulum_kurikulum__id_kurikulum__patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"id_kurikulum","in":"path","required":true,"schema":{"type":"string","title":"Id Kurikulum"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/KurikulumUpdate"}}}},"responses":{"200":{"description":"Data kurikulum yang telah diupdate","content":{"application/json":{"schema":{}}}},"404":{"description":"Tidak ditemukan"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["kurikulum"],"summary":"Detail Kurikulum","description":"Mengambil detail lengkap kurikulum beserta daftar CPL yang terkait","operationId":"detail_kurikulum_kurikulum__id_kurikulum__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"id_kurikulum","in":"path","required":true,"schema":{"type":"string","title":"Id Kurikulum"}}],"responses":{"200":{"description":"Data lengkap kurikulum dengan CPL","content":{"application/json":{"schema":{}}}},"404":{"description":"Tidak ditemukan"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["kurikulum"],"summary":"Hapus Kurikulum","description":"Menghapus kurikulum dari sistem","operationId":"delete_kurikulum_kurikulum__id_kurikulum__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"id_kurikulum","in":"path","required":true,"schema":{"type":"string","title":"Id Kurikulum"}}],"responses":{"204":{"description":"Tidak ada konten (sukses)"},"404":{"description":"Tidak ditemukan"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cpl/{id_kurikulum}":{"post":{"tags":["cpl"],"summary":"Tambah CPL Baru","description":"Menambahkan CPL (Capaian Pembelajaran Lulusan) baru ke kurikulum tertentu","operationId":"create_cpl_cpl__id_kurikulum__post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"id_kurikulum","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Id Kurikulum"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCPL"}}}},"responses":{"201":{"description":"Data CPL yang berhasil ditambahkan","content":{"application/json":{"schema":{}}}},"404":{"description":"Tidak ditemukan"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cpl/{id_kurikulum}/{id_cpl}":{"get":{"tags":["cpl"],"summary":"Detail CPL Lengkap","description":"Mengambil detail lengkap CPL beserta kurikulum, indikator, dan mata kuliah terkait","operationId":"get_detail_cpl_cpl__id_kurikulum___id_cpl__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"id_kurikulum","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Id Kurikulum"}},{"name":"id_cpl","in":"path","required":true,"schema":{"type":"string","title":"Id Cpl"}}],"responses":{"200":{"description":"Data lengkap CPL dengan semua relasinya","content":{"application/json":{"schema":{}}}},"404":{"description":"Tidak ditemukan"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["cpl"],"summary":"Update CPL","description":"Mengupdate deskripsi CPL","operationId":"update_cpl_cpl__id_kurikulum___id_cpl__patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"id_kurikulum","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Id Kurikulum"}},{"name":"id_cpl","in":"path","required":true,"schema":{"type":"string","title":"Id Cpl"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCPL"}}}},"responses":{"200":{"description":"Data CPL yang telah diupdate","content":{"application/json":{"schema":{}}}},"404":{"description":"Tidak ditemukan"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["cpl"],"summary":"Hapus CPL","description":"Menghapus CPL dari kurikulum","operationId":"delete_cpl_cpl__id_kurikulum___id_cpl__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"id_kurikulum","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Id Kurikulum"}},{"name":"id_cpl","in":"path","required":true,"schema":{"type":"string","title":"Id Cpl"}}],"responses":{"204":{"description":"Tidak ada konten (sukses)"},"404":{"description":"Tidak ditemukan"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cpl/kurikulum-aktif":{"get":{"tags":["cpl"],"summary":"Daftar CPL dari Kurikulum Aktif","description":"Mengambil semua CPL yang berasal dari kurikulum dengan status aktif","operationId":"get_cpl_from_active_kurikulum_cpl_kurikulum_aktif_get","responses":{"200":{"description":"Daftar CPL dari kurikulum aktif","content":{"application/json":{"schema":{}}}},"404":{"description":"Tidak ditemukan"}},"security":[{"HTTPBearer":[]}]}},"/indikator/{id_kurikulum}/{id_cpl}":{"post":{"tags":["indikator"],"summary":"Tambah Indikator CPL","description":"Menambahkan indikator baru untuk CPL tertentu dalam kurikulum","operationId":"create_indikator_indikator__id_kurikulum___id_cpl__post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"id_kurikulum","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Id Kurikulum"}},{"name":"id_cpl","in":"path","required":true,"schema":{"type":"string","title":"Id Cpl"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateIndikator"}}}},"responses":{"201":{"description":"Data indikator yang berhasil ditambahkan","content":{"application/json":{"schema":{}}}},"404":{"description":"Tidak ditemukan"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/indikator/{id_kurikulum}/{id_cpl}/{id_indikator}":{"delete":{"tags":["indikator"],"summary":"Hapus Indikator CPL","description":"Menghapus indikator CPL dari sistem","operationId":"deleteIndikator_indikator__id_kurikulum___id_cpl___id_indikator__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"id_kurikulum","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Id Kurikulum"}},{"name":"id_cpl","in":"path","required":true,"schema":{"type":"string","title":"Id Cpl"}},{"name":"id_indikator","in":"path","required":true,"schema":{"type":"string","title":"Id Indikator"}}],"responses":{"204":{"description":"Tidak ada konten (sukses)"},"404":{"description":"Tidak ditemukan"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["indikator"],"summary":"Update Indikator CPL","description":"Mengupdate informasi indikator CPL, termasuk mengubah CPL parent-nya","operationId":"update_indikator_indikator__id_kurikulum___id_cpl___id_indikator__patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"id_kurikulum","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Id Kurikulum"}},{"name":"id_cpl","in":"path","required":true,"schema":{"type":"string","title":"Id Cpl"}},{"name":"id_indikator","in":"path","required":true,"schema":{"type":"string","title":"Id Indikator"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IndikatorCPLUpdate"}}}},"responses":{"200":{"description":"Data indikator yang telah diupdate","content":{"application/json":{"schema":{}}}},"404":{"description":"Tidak ditemukan"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/matkul/":{"get":{"tags":["matkul"],"summary":"Daftar Semua Mata Kuliah","description":"Mengambil daftar semua mata kuliah beserta CPL yang terkait","operationId":"getAllMatkul_matkul__get","responses":{"200":{"description":"Daftar lengkap mata kuliah dengan CPL masing-masing","content":{"application/json":{"schema":{}}}},"404":{"description":"Tidak ditemukan"}},"security":[{"HTTPBearer":[]}]},"post":{"tags":["matkul"],"summary":"Tambah Mata Kuliah Baru","description":"Menambahkan mata kuliah baru beserta relasi dengan CPL (Capaian Pembelajaran Lulusan)","operationId":"inputMatkul_matkul__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/createMatkul"}}},"required":true},"responses":{"201":{"description":"Data mata kuliah dan relasi CPL yang berhasil ditambahkan","content":{"application/json":{"schema":{}}}},"404":{"description":"Tidak ditemukan"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/matkul/{id_matkul}":{"delete":{"tags":["matkul"],"summary":"Hapus Mata Kuliah","description":"Menghapus mata kuliah beserta semua relasi CPL yang terkait","operationId":"deleteMatkul_matkul__id_matkul__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"id_matkul","in":"path","required":true,"schema":{"type":"string","title":"Id Matkul"}}],"responses":{"204":{"description":"Tidak ada konten (sukses)"},"404":{"description":"Tidak ditemukan"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["matkul"],"summary":"Update Mata Kuliah","description":"Mengupdate informasi mata kuliah dan/atau relasi CPL","operationId":"updateMatkul_matkul__id_matkul__patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"id_matkul","in":"path","required":true,"schema":{"type":"string","title":"Id Matkul"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/updateMatkul"}}}},"responses":{"200":{"description":"Data mata kuliah dan relasi CPL yang telah diupdate","content":{"application/json":{"schema":{}}}},"404":{"description":"Tidak ditemukan"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["matkul"],"summary":"Detail Mata Kuliah","description":"Mengambil detail lengkap mata kuliah beserta CPL dan indikator yang terkait","operationId":"getDetailMatkul_matkul__id_matkul__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"id_matkul","in":"path","required":true,"schema":{"type":"string","title":"Id Matkul"}}],"responses":{"200":{"description":"Data lengkap mata kuliah dengan CPL dan indikator","content":{"application/json":{"schema":{}}}},"404":{"description":"Tidak ditemukan"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/cocktails/":{"get":{"tags":["Cocktails"],"summary":"List Cocktails","description":"List cocktails by name","operationId":"list_cocktails_api_cocktails__get","parameters":[{"name":"name","in":"query","required":true,"schema":{"type":"string","title":"Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/cocktails/{cocktail_id}":{"get":{"tags":["Cocktails"],"summary":"Cocktail Detail","description":"Get cocktail detail by ID","operationId":"cocktail_detail_api_cocktails__cocktail_id__get","parameters":[{"name":"cocktail_id","in":"path","required":true,"schema":{"type":"string","title":"Cocktail Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/cocktails/by-letter/{letter}":{"get":{"tags":["Cocktails"],"summary":"Cocktails By Letter","description":"List cocktails by first letter (a-z)","operationId":"cocktails_by_letter_api_cocktails_by_letter__letter__get","parameters":[{"name":"letter","in":"path","required":true,"schema":{"type":"string","title":"Letter"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"CPLInput":{"properties":{"id_kurikulum":{"type":"string","format":"uuid","title":"Id Kurikulum"},"id_cpl":{"type":"string","title":"Id Cpl"}},"type":"object","required":["id_kurikulum","id_cpl"],"title":"CPLInput"},"CreateCPL":{"properties":{"id_cpl":{"type":"string","title":"Id Cpl"},"deskripsi":{"type":"string","title":"Deskripsi"}},"type":"object","required":["id_cpl","deskripsi"],"title":"CreateCPL"},"CreateIndikator":{"properties":{"id_indikator":{"type":"string","title":"Id Indikator"},"deskripsi":{"type":"string","title":"Deskripsi"}},"type":"object","required":["id_indikator","deskripsi"],"title":"CreateIndikator"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"IndikatorCPLUpdate":{"properties":{"deskripsi":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Deskripsi"},"id_cpl":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Id Cpl"}},"type":"object","title":"IndikatorCPLUpdate"},"KurikulumCreate":{"properties":{"nama_kurikulum":{"type":"string","title":"Nama Kurikulum"},"revisi":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Revisi"},"status_kurikulum":{"$ref":"#/components/schemas/StatusEnum","default":"aktif"}},"type":"object","required":["nama_kurikulum"],"title":"KurikulumCreate"},"KurikulumUpdate":{"properties":{"nama_kurikulum":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Nama Kurikulum"},"revisi":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Revisi"},"status_kurikulum":{"anyOf":[{"$ref":"#/components/schemas/StatusEnum"},{"type":"null"}]}},"type":"object","title":"KurikulumUpdate"},"LoginRequest":{"properties":{"user_id":{"type":"string","title":"User Id"},"password":{"type":"string","title":"Password"}},"type":"object","required":["user_id","password"],"title":"LoginRequest"},"RegisterRequest":{"properties":{"user_id":{"type":"string","maxLength":25,"title":"User Id","description":"User ID unik (max 25 karakter)","examples":["dosen001","kadep001"]},"nama":{"type":"string","maxLength":255,"title":"Nama","description":"Nama lengkap user","examples":["Dr. John Doe"]},"password":{"type":"string","minLength":8,"title":"Password","description":"Password minimal 8 karakter","examples":["SecurePass123!"]},"role":{"$ref":"#/components/schemas/RoleEnum","description":"Role user dalam sistem"}},"type":"object","required":["user_id","nama","password","role"],"title":"RegisterRequest","description":"Schema untuk request registrasi user baru"},"RoleEnum":{"type":"string","enum":["kadep","dosen"],"title":"RoleEnum"},"StatusEnum":{"type":"string","enum":["aktif","nonaktif"],"title":"StatusEnum"},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"}},"type":"object","required":["access_token","token_type"],"title":"TokenResponse"},"UpdateCPL":{"properties":{"deskripsi":{"type":"string","title":"Deskripsi"}},"type":"object","required":["deskripsi"],"title":"UpdateCPL"},"UserResponse":{"properties":{"user_id":{"type":"string","title":"User Id"},"nama":{"type":"string","title":"Nama"},"role":{"type":"string","title":"Role"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["user_id","nama","role","created_at","updated_at"],"title":"UserResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"createMatkul":{"properties":{"id_matkul":{"type":"string","title":"Id Matkul"},"mata_kuliah":{"type":"string","title":"Mata Kuliah"},"sks":{"type":"integer","title":"Sks"},"semester":{"type":"integer","title":"Semester"},"cpl_list":{"items":{"$ref":"#/components/schemas/CPLInput"},"type":"array","title":"Cpl List"}},"type":"object","required":["id_matkul","mata_kuliah","sks","semester","cpl_list"],"title":"createMatkul"},"updateMatkul":{"properties":{"mata_kuliah":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Mata Kuliah"},"sks":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Sks"},"semester":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Semester"},"cpl_list":{"anyOf":[{"items":{"$ref":"#/components/schemas/CPLInput"},"type":"array"},{"type":"null"}],"title":"Cpl List"}},"type":"object","title":"updateMatkul"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}